src/lib/verify_keypair.rs:initial impl
